+++ ext/ratatui_ruby/src/lib.rs
@@ -56,7 +56,7 @@ fn init() -> Result<(), Error> {
     buffer_class.define_method("set_string", magnus::method!(buffer::BufferWrapper::set_string, 4))?;
     buffer_class.define_method("area", magnus::method!(buffer::BufferWrapper::area, 0))?;
 
-    m.define_module_function("init_terminal", function!(init_terminal, 0))?;
+    m.define_module_function("_init_terminal", function!(init_terminal, 2))?;
     m.define_module_function("restore_terminal", function!(restore_terminal, 0))?;
     m.define_module_function("draw", function!(draw, 1))?;
     m.define_module_function("_poll_event", function!(events::poll_event, 0))?;